cleanup some general code smells detected by IntelliJ
- refactor fields that can be local
- refactor usage of deprecated API
- fix some unchecked types
- inline some redundant local variables
- replace iterator loops with foreach loops
- Use Java8's Map#computeIfAbsent method
- remove dangling JavaDocs
- remove redundant parent definitions in poms
- use bulk #addAll operations
- fix occurrences of infinite recursion
- fix wrong datatypes in string formattings
- cleanup unused assignments
- simplify stream API use
- remove (some) redundant throws declarations
- make finalizable attributes final

+++ core/src/main/java/net/automatalib/automata/fsa/impl/compact/CompactNFA.java
@@ -76,7 +76,7 @@ public class CompactNFA<I> extends AbstractCompactSimpleNondet<I, Boolean> imple
 
     @Override
     public void setAccepting(Integer state, boolean accepting) {
-        setAccepting(state, accepting);
+        setAccepting(state.intValue(), accepting);
     }
 
     public void setAccepting(int stateId, boolean accepting) {
